Add fp16
Adds the fp16 types, some intrinsics, some std::ops, and a few other methods. Not carefully exercised.

The methods are implemented directly on f16x4 and f16x8. If we are going to support fp16 on AVX-512, then we will want to switch those over to a trait (maybe a subtrait of `Simd`).