Use intrinsics directly rather than wrappers
Put safe wrappers under a feature gate. The impact on compile time is substantial, measured at 0.14 vs 1.17s on M4 in a release build.

The implementations of the method use unsafe directly rather than going through the safe wrappers.

It would be nice to use the safe intrinsics that landed in Rust 1.87, but those depend on the scope being enclosed with `#[target_feature]`, while this approach to multiversioning depends on the level being known at runtime.