Start conversion operations
Adds a `cvt_u32` operation to `f32` vector types. Quite a bit of refactoring to try to reduce duplication, as the number of operation signatures will increase.

Conversion between same-size float and integer types can probably go in the corresponding traits (SimdFloat, etc). It's the narrowing and widening conversions that need to stay in the impl block for the concrete vector types.